Fix error with PHPUnit bootstrapper determining plugin file to load.
The bootstrapper searches the current directory (root folder) for a plugin file to load, and unfortunately `amp-beta-tester.php` is the first one detected and loaded.

This can be bypassed by changing the current directory to one where no plugin file exists, then setting the global `$_plugin_file` variable to the path of the actual plugin file to load.