Sync beta tester version with AMP
Due to the WordPress Plugin Update API using the folder of the main PHP file as an identifier for updates (`amp` in this case), any other plugin file that is placed in the same folder will have the same version number as the AMP plugin.

Setting the plugin version to that of the AMP plugin or greater will solve this problem.